This patch silences almost all of the compiler warnings on Unix.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@3306 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/impexp/xp/ie_exp_RTF.cpp
@@ -607,7 +607,7 @@ void IE_Exp_RTF::_rtf_compute_font_properties(const _rtf_font_info * pfi,
 	
 	*p_sz_font_name = szFontFamily;
 	
-	if ((ff >= 0) && (ff < NrElements(t_ff)))
+	if ((ff >= 0) && (ff < (int)NrElements(t_ff)))
 		*p_sz_rtf_family = t_ff[ff];
 	else
 		*p_sz_rtf_family = t_ff[GR_Font::FF_Unknown];
